requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
argcomplete==3.6.2
blockbuster==1.5.24
boto3==1.38.39
botocore==1.38.39
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cohere==5.15.0
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fasta2a==0.3.1
fastavro==1.11.1
filelock==3.18.0
fonttools==4.58.4
forbiddenfruit==0.1.4
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.21.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
groq==0.28.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
kiwisolver==1.4.8
langchain-core==0.3.65
langgraph==0.4.8
langgraph-api==0.2.56
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.3
langgraph-prebuilt==0.2.2
langgraph-runtime-inmem==0.3.0
langgraph-sdk==0.1.70
langsmith==0.3.45
logfire==3.21.1
logfire-api==3.21.1
markdown-it-py==3.0.0
matplotlib==3.10.3
mcp==1.9.4
mdurl==0.1.2
mistralai==1.8.2
numpy==2.3.0
openai==1.88.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pip==25.1.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.3.1
pydantic-ai-slim==0.3.1
pydantic_core==2.33.2
pydantic-evals==0.3.1
pydantic-graph==0.3.1
pydantic-settings==2.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.47.0
structlog==25.4.0
tenacity==8.5.0
tokenizers==0.21.1
tqdm==4.67.1
truststore==0.10.1
types-requests==2.32.4.20250611
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
zipp==3.23.0
zstandard==0.23.0
